# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Pebble API
  description: API definitions for the Pebble project, shared between Go backend and React frontend.
  version: 0.4.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Dev Server

paths:
  /auth/login:
    post:
      summary: Login via email/password
      tags: [Auth]
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /transactions/{id}/confirm:
    post:
      summary: Confirm a scored transaction to queue penalties
      tags: [Transactions]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Penalties confirmed

  /line-items/{id}/score:
    put:
      summary: Override the AI impulse score for a line item
      tags: [Transactions]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                override_score:
                  type: integer
                  minimum: 0
                  maximum: 100
      responses:
        '200':
          description: Line item updated

  /penalties:
    get:
      summary: List pending penalties
      tags: [Penalties]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, contested]
      responses:
        '200':
          description: Penalty list

  /penalties/{id}/contest:
    post:
      summary: Contest a penalty
      tags: [Penalties]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Status contested

  /penalties/{id}/confirm:
    post:
      summary: Confirm penalty early (before 24h expires)
      tags: [Penalties]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Status confirmed

  /wallet/balance:
    get:
      summary: Wallet balance and pending totals
      tags: [Wallet]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Wallet snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletBalance'

  /wallet/topup:
    post:
      summary: Top up wallet via Razorpay
      tags: [Wallet]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Topup successful

  /wallet/ledger:
    get:
      summary: Wallet transaction ledger
      tags: [Wallet]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Ledger entries

  /users/me:
    get:
      summary: Current user profile with streak and penalty settings
      tags: [Users]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'

  /insights/weekly:
    get:
      summary: Weekly spending digest with top categories and trends
      tags: [Insights]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Weekly digest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyDigest'

  /insights/benchmark:
    get:
      summary: Anonymised cohort impulse benchmark
      tags: [Insights]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Cohort comparison
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenchmarkResult'

  /referrals/me:
    get:
      summary: Referral code and redemption stats
      tags: [Referrals]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Referral programme status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferralStats'

  /referrals/redeem:
    post:
      summary: Redeem a referral code (once per user)
      tags: [Referrals]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        '200':
          description: Code redeemed

  /portfolio:
    get:
      summary: Portfolio breakdown and allocation
      tags: [Investments]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Portfolio summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioResponse'

  /investments:
    get:
      summary: Investment history
      tags: [Investments]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: trigger_type
          in: query
          schema:
            type: string
            enum: [threshold, opportunity, time]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Investment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestmentListResponse'

  /investments/{id}:
    get:
      summary: Single investment detail
      tags: [Investments]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Investment detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Investment'

  /market/signal:
    get:
      summary: Current market opportunity signals (from Redis)
      tags: [Investments]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Live market signals for dashboard indicator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketSignalResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user_id:
          type: string
          format: uuid

    WalletBalance:
      type: object
      properties:
        balance:
          type: number
        pending_total:
          type: number
        invested_total:
          type: number

    Investment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        asset_class:
          type: string
        amount:
          type: number
        units:
          type: number
        nav_at_purchase:
          type: number
        status:
          type: string
        trigger_type:
          type: string
        broker_ref:
          type: string
        created_at:
          type: string
          format: date-time

    PortfolioResponse:
      type: object
      properties:
        total_invested:
          type: number
        equity_value:
          type: number
        gold_value:
          type: number
        bond_value:
          type: number
        gain_pct:
          type: number
        allocation_pct:
          type: object
          additionalProperties:
            type: number

    MarketSignal:
      type: object
      properties:
        asset_class:
          type: string
        indicator:
          type: string
        value:
          type: number
        action:
          type: string
        timestamp:
          type: string
          format: date-time

    MarketSignalResponse:
      type: object
      properties:
        signals:
          type: array
          items:
            $ref: '#/components/schemas/MarketSignal'
        composite_score:
          type: number
        updated_at:
          type: string

    InvestmentListResponse:
      type: object
      properties:
        investments:
          type: array
          items:
            $ref: '#/components/schemas/Investment'
        total:
          type: integer

    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        phone:
          type: string
        risk_profile:
          type: string
        penalty_rate:
          type: number
        effective_penalty_rate:
          type: number
        penalty_threshold:
          type: integer
        invest_threshold:
          type: number
        consent_hours:
          type: integer
        streak_count:
          type: integer
        streak_last_updated:
          type: string
          format: date-time
          nullable: true
        streak_discount_pct:
          type: number

    WeeklyDigest:
      type: object
      properties:
        week_start:
          type: string
          format: date-time
        week_end:
          type: string
          format: date-time
        total_spend:
          type: number
        impulse_pct:
          type: number
        avg_impulse_score:
          type: number
        top_categories:
          type: array
          items:
            $ref: '#/components/schemas/CategorySpend'
        trend_vs_last_week_pct:
          type: number

    CategorySpend:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        pct:
          type: number

    ReferralStats:
      type: object
      properties:
        code:
          type: string
        redemption_count:
          type: integer
        discount_pct:
          type: number

    BenchmarkResult:
      type: object
      properties:
        user_impulse_pct:
          type: number
        cohort_impulse_pct:
          type: number
        saved_vs_cohort_pct:
          type: number
        cohort_label:
          type: string
        sample_size:
          type: integer